Module aws_sdk_iotanalytics::error

Modules

internal_failure_exception

See InternalFailureException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

Structs

BatchPutMessageError

CancelPipelineReprocessingError

CreateChannelError

CreateDatasetContentError

CreateDatasetError

CreateDatastoreError

CreatePipelineError

DeleteChannelError

DeleteDatasetContentError

DeleteDatasetError

DeleteDatastoreError

DeletePipelineError

DescribeChannelError

DescribeDatasetError

DescribeDatastoreError

DescribeLoggingOptionsError

DescribePipelineError

GetDatasetContentError

InternalFailureException

There was an internal failure.

InvalidRequestException

The request was not valid.

LimitExceededException

The command caused an internal limit to be exceeded.

ListChannelsError

ListDatasetContentsError

ListDatasetsError

ListDatastoresError

ListPipelinesError

ListTagsForResourceError

PutLoggingOptionsError

ResourceAlreadyExistsException

A resource with the same name already exists.

ResourceNotFoundException

A resource with the specified name could not be found.

RunPipelineActivityError

SampleChannelDataError

ServiceUnavailableException

The service is temporarily unavailable.

StartPipelineReprocessingError

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateChannelError

UpdateDatasetError

UpdateDatastoreError

UpdatePipelineError

Enums

BatchPutMessageErrorKind

CancelPipelineReprocessingErrorKind

CreateChannelErrorKind

CreateDatasetContentErrorKind

CreateDatasetErrorKind

CreateDatastoreErrorKind

CreatePipelineErrorKind

DeleteChannelErrorKind

DeleteDatasetContentErrorKind

DeleteDatasetErrorKind

DeleteDatastoreErrorKind

DeletePipelineErrorKind

DescribeChannelErrorKind

DescribeDatasetErrorKind

DescribeDatastoreErrorKind

DescribeLoggingOptionsErrorKind

DescribePipelineErrorKind

GetDatasetContentErrorKind

ListChannelsErrorKind

ListDatasetContentsErrorKind

ListDatasetsErrorKind

ListDatastoresErrorKind

ListPipelinesErrorKind

ListTagsForResourceErrorKind

PutLoggingOptionsErrorKind

RunPipelineActivityErrorKind

SampleChannelDataErrorKind

StartPipelineReprocessingErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateChannelErrorKind

UpdateDatasetErrorKind

UpdateDatastoreErrorKind

UpdatePipelineErrorKind