Module aws_sdk_iotanalytics::model

Modules

add_attributes_activity

See AddAttributesActivity

batch_put_message_error_entry

See BatchPutMessageErrorEntry

channel

See Channel

channel_activity

See ChannelActivity

channel_messages

See ChannelMessages

channel_statistics

See ChannelStatistics

channel_storage

See ChannelStorage

channel_storage_summary

See ChannelStorageSummary

channel_summary

See ChannelSummary

column

See Column

container_dataset_action

See ContainerDatasetAction

customer_managed_channel_s3_storage

See CustomerManagedChannelS3Storage

customer_managed_channel_s3_storage_summary

See CustomerManagedChannelS3StorageSummary

customer_managed_datastore_s3_storage

See CustomerManagedDatastoreS3Storage

customer_managed_datastore_s3_storage_summary

See CustomerManagedDatastoreS3StorageSummary

dataset

See Dataset

dataset_action

See DatasetAction

dataset_action_summary

See DatasetActionSummary

dataset_content_delivery_destination

See DatasetContentDeliveryDestination

dataset_content_delivery_rule

See DatasetContentDeliveryRule

dataset_content_status

See DatasetContentStatus

dataset_content_summary

See DatasetContentSummary

dataset_content_version_value

See DatasetContentVersionValue

dataset_entry

See DatasetEntry

dataset_summary

See DatasetSummary

dataset_trigger

See DatasetTrigger

datastore

See Datastore

datastore_activity

See DatastoreActivity

datastore_iot_site_wise_multi_layer_storage

See DatastoreIotSiteWiseMultiLayerStorage

datastore_iot_site_wise_multi_layer_storage_summary

See DatastoreIotSiteWiseMultiLayerStorageSummary

datastore_partition

See DatastorePartition

datastore_partitions

See DatastorePartitions

datastore_statistics

See DatastoreStatistics

datastore_storage_summary

See DatastoreStorageSummary

datastore_summary

See DatastoreSummary

delta_time

See DeltaTime

delta_time_session_window_configuration

See DeltaTimeSessionWindowConfiguration

device_registry_enrich_activity

See DeviceRegistryEnrichActivity

device_shadow_enrich_activity

See DeviceShadowEnrichActivity

estimated_resource_size

See EstimatedResourceSize

file_format_configuration

See FileFormatConfiguration

filter_activity

See FilterActivity

glue_configuration

See GlueConfiguration

iot_events_destination_configuration

See IotEventsDestinationConfiguration

iot_site_wise_customer_managed_datastore_s3_storage

See IotSiteWiseCustomerManagedDatastoreS3Storage

iot_site_wise_customer_managed_datastore_s3_storage_summary

See IotSiteWiseCustomerManagedDatastoreS3StorageSummary

json_configuration

See JsonConfiguration

lambda_activity

See LambdaActivity

late_data_rule

See LateDataRule

late_data_rule_configuration

See LateDataRuleConfiguration

logging_options

See LoggingOptions

math_activity

See MathActivity

message

See Message

output_file_uri_value

See OutputFileUriValue

parquet_configuration

See ParquetConfiguration

partition

See Partition

pipeline

See Pipeline

pipeline_activity

See PipelineActivity

pipeline_summary

See PipelineSummary

query_filter

See QueryFilter

remove_attributes_activity

See RemoveAttributesActivity

reprocessing_summary

See ReprocessingSummary

resource_configuration

See ResourceConfiguration

retention_period

See RetentionPeriod

s3_destination_configuration

See S3DestinationConfiguration

schedule

See Schedule

schema_definition

See SchemaDefinition

select_attributes_activity

See SelectAttributesActivity

service_managed_channel_s3_storage

See ServiceManagedChannelS3Storage

service_managed_channel_s3_storage_summary

See ServiceManagedChannelS3StorageSummary

service_managed_datastore_s3_storage

See ServiceManagedDatastoreS3Storage

service_managed_datastore_s3_storage_summary

See ServiceManagedDatastoreS3StorageSummary

sql_query_dataset_action

See SqlQueryDatasetAction

tag

See Tag

timestamp_partition

See TimestampPartition

triggering_dataset

See TriggeringDataset

variable

See Variable

versioning_configuration

See VersioningConfiguration

Structs

AddAttributesActivity

An activity that adds other attributes based on existing attributes in the message.

BatchPutMessageErrorEntry

Contains informations about errors.

Channel

A collection of data from an MQTT topic. Channels archive the raw, unprocessed messages

ChannelActivity

The activity that determines the source of the messages to be processed.

ChannelMessages

Specifies one or more sets of channel messages.

ChannelStatistics

Statistics information about the channel.

ChannelStorage

Where channel data is stored. You may choose one of serviceManagedS3,

ChannelStorageSummary

Where channel data is stored.

ChannelSummary

A summary of information about a channel.

Column

Contains information about a column that stores your data.

ContainerDatasetAction

Information required to run the containerAction to produce dataset

CustomerManagedChannelS3Storage

Used to store channel data in an S3 bucket that you manage. If customer-managed storage is

CustomerManagedChannelS3StorageSummary

Used to store channel data in an S3 bucket that you manage.

CustomerManagedDatastoreS3Storage

S3-customer-managed; When you choose customer-managed storage, the retentionPeriod parameter is ignored. You can't change the choice of Amazon S3 storage after your data store is created.

CustomerManagedDatastoreS3StorageSummary

Contains information about the data store that you manage.

Dataset

Information about a dataset.

DatasetAction

A DatasetAction object that specifies how dataset contents are automatically

DatasetActionSummary

Information about the action that automatically creates the dataset's contents.

DatasetContentDeliveryDestination

The destination to which dataset contents are delivered.

DatasetContentDeliveryRule

When dataset contents are created, they are delivered to destination specified

DatasetContentStatus

The state of the dataset contents and the reason they are in this state.

DatasetContentSummary

Summary information about dataset contents.

DatasetContentVersionValue

The dataset whose latest contents are used as input to the notebook or application.

DatasetEntry

The reference to a dataset entry.

DatasetSummary

A summary of information about a dataset.

DatasetTrigger

The DatasetTrigger that specifies when the dataset is automatically

Datastore

Information about a data store.

DatastoreActivity

The datastore activity that specifies where to store the processed data.

DatastoreIotSiteWiseMultiLayerStorage

Used to store data used by IoT SiteWise in an Amazon S3 bucket that you manage. You can't change the choice of Amazon S3 storage after your data store is created.

DatastoreIotSiteWiseMultiLayerStorageSummary

Contains information about the data store that you manage, which stores data used by IoT SiteWise.

DatastorePartition

A single dimension to partition a data store. The dimension must be an

DatastorePartitions

Contains information about the partition dimensions in a data store.

DatastoreStatistics

Statistical information about the data store.

DatastoreStorageSummary

Contains information about your data store.

DatastoreSummary

A summary of information about a data store.

DeltaTime

Used to limit data to that which has arrived since the last execution of the

DeltaTimeSessionWindowConfiguration

A structure that contains the configuration information of a delta time session

DeviceRegistryEnrichActivity

An activity that adds data from the IoT device registry to your message.

DeviceShadowEnrichActivity

An activity that adds information from the IoT Device Shadow service to a

EstimatedResourceSize

The estimated size of the resource.

FileFormatConfiguration

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

FilterActivity

An activity that filters a message based on its attributes.

GlueConfiguration

Configuration information for coordination with Glue, a fully managed extract, transform

IotEventsDestinationConfiguration

Configuration information for delivery of dataset contents to IoT Events.

IotSiteWiseCustomerManagedDatastoreS3Storage

Used to store data used by IoT SiteWise in an Amazon S3 bucket that you manage. You can't change the choice of Amazon S3 storage after your data store is created.

IotSiteWiseCustomerManagedDatastoreS3StorageSummary

Contains information about the data store that you manage, which stores data used by IoT SiteWise.

JsonConfiguration

Contains the configuration information of the JSON format.

LambdaActivity

An activity that runs a Lambda function to modify the message.

LateDataRule

A structure that contains the name and configuration information of a late data

LateDataRuleConfiguration

The information needed to configure a delta time session window.

LoggingOptions

Information about logging options.

MathActivity

An activity that computes an arithmetic expression using the message's attributes.

Message

Information about a message.

OutputFileUriValue

The value of the variable as a structure that specifies an output file URI.

ParquetConfiguration

Contains the configuration information of the Parquet format.

Partition

A partition dimension defined by an attribute.

Pipeline

Contains information about a pipeline.

PipelineActivity

An activity that performs a transformation on a message.

PipelineSummary

A summary of information about a pipeline.

QueryFilter

Information that is used to filter message data, to segregate it according to the

RemoveAttributesActivity

An activity that removes attributes from a message.

ReprocessingSummary

Information about pipeline reprocessing.

ResourceConfiguration

The configuration of the resource used to execute the containerAction.

RetentionPeriod

How long, in days, message data is kept.

S3DestinationConfiguration

Configuration information for delivery of dataset contents to Amazon Simple Storage Service (Amazon S3).

Schedule

The schedule for when to trigger an update.

SchemaDefinition

Information needed to define a schema.

SelectAttributesActivity

Used to create a new message using only the specified attributes from the original

ServiceManagedChannelS3Storage

Used to store channel data in an S3 bucket managed by IoT Analytics. You can't change the choice

ServiceManagedChannelS3StorageSummary

Used to store channel data in an S3 bucket managed by IoT Analytics.

ServiceManagedDatastoreS3Storage

Used to store data in an Amazon S3 bucket managed by IoT Analytics. You can't change the choice of Amazon S3 storage after your data store is created.

ServiceManagedDatastoreS3StorageSummary

Contains information about the data store that is managed by IoT Analytics.

SqlQueryDatasetAction

The SQL query to modify the message.

Tag

A set of key-value pairs that are used to manage the resource.

TimestampPartition

A partition dimension defined by a timestamp attribute.

TriggeringDataset

Information about the dataset whose content generation triggers the new dataset content

Variable

An instance of a variable to be passed to the containerAction execution. Each

VersioningConfiguration

Information about the versioning of dataset contents.

Enums

ChannelStatus

ComputeType

DatasetActionType

DatasetContentState

DatasetStatus

DatastoreStatus

DatastoreStorage

FileFormatType

LoggingLevel

ReprocessingStatus