Module aws_sdk_iotanalytics::operation

Structs

BatchPutMessage

Sends messages to a channel.

CancelPipelineReprocessing

Cancels the reprocessing of data through the pipeline.

CreateChannel

Used to create a channel. A channel collects data from an MQTT topic and archives the raw,

CreateDataset

Used to create a dataset. A dataset stores data retrieved from a data store by applying a

CreateDatasetContent

Creates the content of a dataset by applying a queryAction (a SQL query) or a

CreateDatastore

Creates a data store, which is a repository for messages.

CreatePipeline

Creates a pipeline. A pipeline consumes messages from a channel and allows you to process

DeleteChannel

Deletes the specified channel.

DeleteDataset

Deletes the specified dataset.

DeleteDatasetContent

Deletes the content of the specified dataset.

DeleteDatastore

Deletes the specified data store.

DeletePipeline

Deletes the specified pipeline.

DescribeChannel

Retrieves information about a channel.

DescribeDataset

Retrieves information about a dataset.

DescribeDatastore

Retrieves information about a data store.

DescribeLoggingOptions

Retrieves the current settings of the IoT Analytics logging options.

DescribePipeline

Retrieves information about a pipeline.

GetDatasetContent

Retrieves the contents of a dataset as presigned URIs.

ListChannels

Retrieves a list of channels.

ListDatasetContents

Lists information about dataset contents that have been created.

ListDatasets

Retrieves information about datasets.

ListDatastores

Retrieves a list of data stores.

ListPipelines

Retrieves a list of pipelines.

ListTagsForResource

Lists the tags (metadata) that you have assigned to the resource.

PutLoggingOptions

Sets or updates the IoT Analytics logging options.

RunPipelineActivity

Simulates the results of running a pipeline activity on a message payload.

SampleChannelData

Retrieves a sample of messages from the specified channel ingested during the specified

StartPipelineReprocessing

Starts the reprocessing of raw message data through the pipeline.

TagResource

Adds to or modifies the tags of the given resource. Tags are metadata that can be used to

UntagResource

Removes the given tags (metadata) from the resource.

UpdateChannel

Used to update the settings of a channel.

UpdateDataset

Updates the settings of a dataset.

UpdateDatastore

Used to update the settings of a data store.

UpdatePipeline

Updates the settings of a pipeline. You must specify both a channel and a